#791 Moved test domain entities from src/test to src/main in order to resolve maven compilation errors.
