+++ platform-dao/src/test/java/ua/com/fielden/platform/sample/domain/EntityOneDao.java
@@ -1,10 +1,11 @@
-package ua.com.fielden.platform.entity.union;
+package ua.com.fielden.platform.sample.domain;
 
 import com.google.inject.Inject;
 
 import ua.com.fielden.platform.dao.CommonEntityDao;
 import ua.com.fielden.platform.entity.annotation.EntityType;
 import ua.com.fielden.platform.entity.query.IFilter;
+import ua.com.fielden.platform.sample.domain.IEntityOne;
 
 /**
  * DAO implementation for companion object {@link IEntityOne}.
