+++ platform-pojo-bl/src/test/java/ua/com/fielden/platform/entity/union/UnionEntityWithKindOneError.java
@@ -4,6 +4,7 @@ import ua.com.fielden.platform.entity.AbstractUnionEntity;
 import ua.com.fielden.platform.entity.annotation.IsProperty;
 import ua.com.fielden.platform.entity.annotation.Observable;
 import ua.com.fielden.platform.entity.annotation.Title;
+import ua.com.fielden.platform.sample.domain.EntityOne;
 
 public class UnionEntityWithKindOneError extends AbstractUnionEntity {
     @IsProperty
